Create test for new components and hooks
