+++ x-pack/plugins/security_solution/public/detection_engine/rule_details_ui/pages/rule_details/__mocks__/rule_details_context.tsx
@@ -7,7 +7,7 @@
 
 import React from 'react';
 import type { RuleDetailsContextType } from '../rule_details_context';
-import { RuleDetailTabs } from '..';
+import { RuleDetailTabs } from '../use_rule_details_tabs';
 
 export const useRuleDetailsContextMock = {
   create: (): jest.Mocked<RuleDetailsContextType> => ({
