+++ x-pack/plugins/security_solution/public/detection_engine/rule_details_ui/pages/rule_details/use_redirect_legacy_url.ts
@@ -10,7 +10,7 @@ import { i18n as i18nTranslate } from '@kbn/i18n';
 import type { SpacesApi } from '@kbn/spaces-plugin/public';
 import type { Rule } from '../../../rule_management/logic';
 
-interface UseLegacyUrlRedirectParams {
+export interface UseLegacyUrlRedirectParams {
   rule: Rule | null;
   spacesApi?: SpacesApi;
 }
